requirements.txt
# Google APIs
google-api-python-client>=2.108.0,<3.0.0
google-auth-oauthlib>=1.1.0,<2.0.0
google-auth>=2.23.0,<3.0.0

# Text-to-Speech
gTTS>=2.3.2,<3.0.0

# Image Processing
Pillow>=10.0.0,<11.0.0

# HTTP Requests
requests>=2.31.0,<3.0.0

# Type hints support for older Python versions
typing-extensions>=4.0.0; python_version<"3.8" 